refactor: better naming and better file structure
